# public/Swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Bilemo",
        "description": "This is the Bilemo server.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Api BileMo"
        }
    ],
    "paths": {
        "/mobiles": {
            "get": {
                "tags": [
                    "Mobiles"
                ],
                "summary": "Return a list of phones ressource",
                "operationId": "32acfc985d6def6ddc382e6d6df6ee19",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of the mobile phones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "JWT Token not found or expired"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/mobiles/{id}": {
            "get": {
                "tags": [
                    "Mobiles"
                ],
                "summary": "Return phone details",
                "operationId": "358dda1353c9977c9bb71fa5de88224a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the mobile phone to get",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details for one mobile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "JWT Token not found or expired"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/login_check": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Return a JWT token for authentification",
                "operationId": "9056e654130817d5830fe3912b502e81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "The username of the client.",
                                        "type": "string",
                                        "example": "yourUsername"
                                    },
                                    "password": {
                                        "description": "Password of the client.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "yourPassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticate token JWT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Return a list of users for the client",
                "operationId": "188bef79089c193d56b951242835f198",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of the mobile phones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "JWT Token not found or expired"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "36c4aa0173e0bb6d43e416d166ed7b46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "The username of the user.",
                                        "type": "string",
                                        "example": "usernameUser"
                                    },
                                    "password": {
                                        "description": "Password of the user.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "passwordUser123"
                                    },
                                    "firstname": {
                                        "description": "Firstname of the new user.",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastname": {
                                        "description": "Lastname of the new user.",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "description": "Email address of the new user.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john_doe@gmail.fr"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creation of an user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "description": "ID of the user created"
                                        }
                                    }
                                }
                            }
                        },
                        "links" : {
                            "GetUserByUserId": {
                                "description": "The ID value returned can be used as the ID parameter in GET users/{id}",
                                "operationId": "getUser",
                                "parameters": {
                                    "id": "$response.body#/id"
                                }
                            },
                            "DeleteUserByUserId": {
                                "description": "The ID value returned can be used as the ID parameter in DELETE users/{id}",
                                "operationId": "deleteUser",
                                "parameters": {
                                    "id": "$response.body#/id"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "JWT Token not found or expired"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Return user client details",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details for one user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "JWT Token not found or expired"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete an user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The user has been deleted"
                    },
                    "401": {
                        "description": "JWT Token not found or expired"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "description": "The hashed password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "brand": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "json"
                    },
                    "password": {
                        "description": "The hashed password",
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "dateCreate": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}